+++ src/main/resources/static/x-admin/MyJs/view_list.js
@@ -36,12 +36,16 @@ function showUsersTable(pageInfo){
             +'<td>'+replaceNull(users[i].pictureNum)+'</td>'
             +'<td>'+replaceNull(users[i].commentNum)+'</td>'
             +'<td>'+replaceNull(users[i].createtime)+'</td>'
-            +'<td>'
-            +'<a href=javascript:void(0) title="点击修改景点" onclick="updateUser('+users[i].id+')"><i class="layui-icon">&#xe642;</i></a>'
+            +'<td>';
+        if("undefined" != typeof admin){
+            tr+='<a href=javascript:void(0) title="点击修改景点" onclick="updateUser('+users[i].id+')"><i class="layui-icon">&#xe642;</i></a>'
             +'<a href=javascript:void(0) title="点击删除该景点" onclick="deleteUser('+users[i].id+')"><i class="layui-icon">&#xe640;</i></a>'
             +'<a href=javascript:void(0) title="点击上传景点图片" onclick="showPicture('+users[i].id+')"><i class="layui-icon">&#xe681;</i></a>'
             +'<a href=javascript:void(0) title="点击查看详情" onclick="queryViewDetails('+users[i].id+')"><i class="layui-icon">&#xe615;</i></a>'
-            +'</td></tr>'
+            }else{
+            	tr+='-- --'
+            }
+            tr +='</td></tr>'
         $("#memberTbody").append(tr);
     }
